// admin-clients/libs/cpanel/organizations/entity-users/data-access/src/lib/models/entity-user-permissions.enum.ts
export enum EntityUserPermissions {
    /* Permisos rol usuario operadora */
    automaticSales = 'AUTOMATIC_SALES',
    /* Permisos rol usuario de taquilla */
    invite = 'INVITE',
    refund = 'REFUND',
    issue = 'ISSUE',
    /* Permisos rol integración de canales */
    secondMarket = 'SECOND_MARKET',
    /* Permisos BI */
    advanced = 'BI_ADVANCED',
    basic = 'BI_BASIC',
    mobile = 'BI_MOBILE',
    biInherit = 'BI_INHERIT',
    impersonation = 'BI_IMPERSONATION'
}